backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
transformers==4.35.2
torch>=2.2.0
torchvision>=0.17.0
pillow==10.1.0
numpy>=1.26.0
geopy==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
python-dotenv==1.0.0
watsonx-ai==0.1.0
ibm-watson==7.1.0
